requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
pypdf
chromadb
python-dotenv